:sparkles: Stage 7 : Add semantics for class methods call
Semantic and syntax checking for class methods being called in the main function has been added. The function calls are of the form:

id.func_name(args)

where 'id' is a class variable, 'func_name' is a class method and 'args' is the list of arguments.

Checks whether data-type returned by the member function is same as of the RHS variable in-case of an assignment statement.

Checks whether the right arguments are passed to the member function.

Checks whether a valid member function is called and not the member field of the class.

classTable.c and classTable.h
+ Global Variable currentCDeclType

GSTable.c and GSTable.h
+ GST entries now has classTablePtr field
+ Pretty-print GST

memberFuncList.c and memberFuncList.h
+ verifyClassFuncArgs()

typeTable.c and typeTable.h
+ FLLookUp() now searches for fields in both type table and class table

ast.c and ast.h
+ classTablePtr field added to ASTNode

ast.l
+ self keyword added

ast.y
+ Grammar for fieldFunction added

FieldFunction	        :	SELF '.' ID '(' ArgList ')'
				|	Field '(' ArgList ')'

+ StructField renamed to Field